Time: 1 ms (77.16%), Space: 42.6 MB (54.43%) - LeetHub

+++ 50-powx-n/50-powx-n.java
@@ -17,7 +17,7 @@ class Solution {
         }
         
         
-        if(n<0) return 1/ans;
+        if(n<0) return (double)1/(double)ans;
         else return ans;
